check type works now change atoi to long

+++ includes/libft/ft_atoi.c
@@ -6,17 +6,17 @@
 /*   By: skunert <skunert@student.42heilbronn.de    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/14 16:50:23 by skunert           #+#    #+#             */
-/*   Updated: 2023/04/25 17:31:40 by skunert          ###   ########.fr       */
+/*   Updated: 2023/04/26 16:03:06 by skunert          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
-	int	i;
-	int	sign;
-	int	res;
+	int		i;
+	int		sign;
+	long	res;
 
 	i = 0;
 	sign = 1;
